requirements.txt
-i https://pypi.org/simple

mysql-connector-python==8.0.33
python-dotenv==0.21.1
sqlalchemy==2.0.20
